refactor: move pre-localization logic to  localization module
